requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
cryptography==1.8.1
cycler==0.10.0
decorator==4.0.11
enum34==1.1.6
functools32==3.2.3.post2
idna==2.5
ipaddress==1.0.18
Lasagne==0.1
matplotlib==2.0.0
networkx==1.11
numpy==1.12.1
olefile==0.44
packaging==16.8
paramiko==2.1.2
pbr==3.0.0
Pillow==4.1.0
pyasn1==0.2.3
pycparser==2.17
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
scikit-image==0.13.0
scipy==0.19.0
six==1.10.0
SQLAlchemy==1.1.9
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
subprocess32==3.2.7
Tempita==0.5.2
Theano==0.9.0
